reworked github_panel.py (1/2)
